moved functions to util
